% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docstring.R
\name{docstring_to_roxygen}
\alias{docstring_to_roxygen}
\title{Attempts to pull out the docstring and turn them into valid
roxygen style comments formated in a way that it should
be able to be written directly to a file}
\usage{
docstring_to_roxygen(fun, funname = as.character(substitute(fun)))
}
\description{
Attempts to pull out the docstring and turn them into valid
roxygen style comments formated in a way that it should
be able to be written directly to a file
}
