% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docstring.R
\name{docstring}
\alias{docstring}
\title{Display a docstring}
\usage{
docstring(fun, default_title = "Title not detected", warnings = TRUE)
}
\arguments{
\item{fun}{The function that has the docstring you would like to display}

\item{default_title}{The title you would like to display if no title is detected
in the docstring itself}

\item{warnings}{logical Whether you want warning messages displayed in the console}
}
\description{
Display a docstring using R's built in help file viewer.
}
